Combine woo-identifiers and SKU to reduce complexity
